remove full screen dialog
