--- src/components/FileManagementDialog/index.ts
@@ -1,1 +1,0 @@
-export { FileManagementDialog } from "./FileManagementDialog";
