--- src/components/FullScreenDialog/index.ts
@@ -1,3 +1,0 @@
-export * from "./FullScreenDialog";
-export * from "./FullScreenDialogContent";
-export * from "./FullScreenDialogTitle";
